Implement Form Submission
Once the form has a name and is submitted, the submit event
will call a function to emit project added so that the form is added
to the sidebar with the correct information. Then call
funtion to remove the form.

Move color circles to the left to prevent
them being effected by long project names