Separate render() Function
Instead of inside, reOrderProjects() will handle
the ordering of the projects, leaving only rendering to render.

renderSideBar is for when the user adds below or above on
the project home. Previously the project was added to
the end and only rendered when the user returned to the home.
Now it be in the correct spot when added.