Update Todo Valitation
Before there was only validation on
the taskName since that was all that was required.

Now everything is required so follow the same
procedure for the other elements as well.

And call a function to check if they are
all valid whenever any of them are changed.

When Invalid Changed the background color to red.